Added first version with scrolling area working properly
- Created the remaining layouts
- Refined to accomodate nested layouts
- Deleted the old interface_gui_layout.ui
- Added the new interface_gui_v2.0.ui with the new layouts

in the interface there was only one scroll area, now there are two:
one for the settings and one for the plots. The settings area
now expands with a minimum size of 370x700, and the plots area has a
minimum of 745x600. These values were chosen arbirarily,
but they seem to work well. Needs more testing to see if they work well
on other machines.

Having only one scroll area was causing problems when resizing
the window. When the window was resized it clipped the settings.
With two it is possible to expand the plots area without
clipping adjacent areas.

For now i also edited the interface_visual.py to call the new interface